pyproject.toml
[build-system]
requires = ["setuptools>=67", "setuptools_scm>=7", "wheel"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
omit = ["./tests/*", "*/worker-base/*"]

[tool.pytest.ini_options]
addopts = "--tb=short --cov --cov-report=html"
DJANGO_SETTINGS_MODULE = "tests.settings"
norecursedirs = [".git", ".vscode", "templates", "htmlcov"]

[tool.ruff]
extend-select = ["I"]

[tool.ruff.format]
preview = true

[tool.setuptools]
packages = ["wagtailgridder"]

[tool.setuptools_scm]
write_to = "wagtailgridder/_version.py"

[project]
name = "wagtailgridder"
authors = [{name = "Timothy Allen", email = "tallen@wharton.upenn.edu"},]
description = "Dynamic, interactive grid layout for the Django CMS Wagtail."
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.7"
keywords = ["wagtail", "cms", "gridder", "theme", "dynamic"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 4",
    "Framework :: Wagtail :: 5",
    "Framework :: Wagtail :: 6",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "wagtail>=4.1",
    "wagtail-modeladmin",
]

[project.optional-dependencies]
dev = [
    "django-coverage-plugin",
    "ipython",
    "pytest-django",
    "pytest-cov",
    "ruff",
]

[project.urls]
"Homepage" = "https://github.com/wharton/wagtailgridder"
"Repository" = "https://github.com/wharton/wagtailgridder"
"Documentation" = "https://github.com/wharton/wagtailgridder"
